Fix bug causing incorrect currency conversion
This commit fixes a bug that was causing the currency conversion to be incorrect in the Rates component. The bug was caused by a mistake in the calculation of the exchange rate.

To fix the bug, I updated the calculation to use the correct formula and added unit tests to ensure that the calculation is correct.

This fix resolves issue #123 and improves the accuracy of the currency conversion in the Rates component.